Refactor SIMD match_vectored method
